requirements.txt
absl-py==2.1.0
accelerate==1.10.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
airportsdata==20250224
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.13.3
bitsandbytes==0.47.0
bleach==6.2.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cut-cross-entropy==25.1.1
cycler==0.12.1
cymem==2.0.11
datasets==3.6.0
debugpy==1.8.13
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
diffusers==0.35.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.7
distro==1.9.0
docopt==0.6.2
docstring-parser==0.17.0
einops==0.8.1
einops-exts==0.0.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
et-xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.12.2
fire==0.7.0
flash-attn==2.8.2
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
funcy==2.0
gast==0.6.0
gdown==5.2.0
genson==1.3.0
google-pasta==0.2.0
gradio==5.21.0
gradio-client==1.7.2
groovy==0.1.2
grpcio==1.73.0
h11==0.14.0
h5py==3.14.0
hf-transfer==0.1.9
hf-xet==1.1.9
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.34.4
identify==2.6.9
idna==3.10
imageio==2.37.0
importlib-metadata==8.7.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
iso3166==2.1.1
isoduration==20.11.0
janus @ git+https://github.com/deepseek-ai/Janus.git@1daa72fa409002d40931bd7b36a9280362469ead
jedi==0.19.2
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
keras==3.11.3
kiwisolver==1.4.8
kornia==0.8.1
kornia-rs==0.1.9
langcodes==3.5.0
language-data==1.3.0
lark==1.2.2
libclang==18.1.1
line-profiler==4.2.0
lmdeploy==0.7.2.post1
marisa-trie==1.2.1
markdown==3.8
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral-common==1.8.5
mistral-inference==1.6.0
mistune==3.1.3
ml-dtypes==0.5.3
mmengine-lite==0.10.7
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.12
namex==0.1.0
narwhals==2.5.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nlpaug==1.1.11
nltk==3.9.1
nodeenv==1.9.1
notebook==7.3.3
notebook-shim==0.2.4
num2words==0.5.14
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai==1.66.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.4.0
optree==0.17.0
orjson==3.10.15
outlines==0.2.1
outlines-core==0.1.26
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.17.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==3.9.1
portalocker==3.1.1
pre-commit==4.2.0
preshed==3.0.9
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-extra-types==2.10.5
pydub==0.25.1
pygments==2.19.2
pynvml==12.0.0
pyparsing==3.2.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
ray==2.44.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.1.0
rouge-score==0.1.2
rpds-py==0.23.1
ruff==0.11.0
safehttpx==0.1.6
safetensors==0.6.2
SciencePlots==2.1.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
simple-parsing==0.1.7
simsimd==6.2.1
six==1.17.0
sklearn==0.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-pkuseg==1.0.0
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.1
stringzilla==3.12.3
sty==1.0.6
svgwrite==1.4.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==2.5.0
terminado==0.18.1
tf-keras==2.20.1
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
timeout-decorator==0.5.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.4
tomli==2.2.1
tomlkit==0.13.2
torch==2.8.0
torchao==0.13.0
torchvision==0.23.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
transformers-stream-generator==0.0.5
triton==3.4.0
trl==0.22.2
typeguard==4.4.4
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-extensions==4.15.0
tyro==0.9.31
tzdata==2025.2
unsloth==2025.8.10
unsloth-zoo==2025.8.9
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.34.0
validators==0.34.0
virtualenv==20.24.1
wasabi==1.1.3
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
xformers==0.0.32.post2
XlsxWriter==3.2.2
xxhash==3.5.0
yapf==0.43.0
yarg==0.1.9
yarl==1.20.1
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.8.0/zh_core_web_sm-3.8.0-py3-none-any.whl
zipp==3.23.0
